+++ packages/eslint-plugin-smarthr/rules/a11y-prohibit-checkbox-or-radio-in-table-cell/index.js
@@ -1,12 +1,3 @@
-const findClosestThFromAncestor = (node) => {
-  if (node.type === 'JSXElement' && node.openingElement.name.name === 'Th') {
-    return node
-  }
-  if (node.parent) {
-    return findClosestThFromAncestor(node.parent)
-  }
-}
-
 /**
  * @type {import('@typescript-eslint/utils').TSESLint.RuleModule<''>}
  */
@@ -20,10 +11,8 @@ module.exports = {
     },
   },
   create(context) {
-    const sourceCode = context.sourceCode
-
     return {
-      'JSXElement[openingElement.name.name=/Td$/] JSXElement[openingElement.name.name=/Check(b|B)ox$/]': (node) => {
+      'JSXElement[openingElement.name.name=/Td$/] JSXElement[openingElement.name.name=/Check(b|B)ox$/][children.length=0]': (node) => {
         context.report({
           node,
           messageId: 'default',
@@ -34,7 +23,7 @@ module.exports = {
           },
         })
       },
-      'JSXElement[openingElement.name.name=/Td$/] JSXElement[openingElement.name.name=/RadioButton$/]': (node) => {
+      'JSXElement[openingElement.name.name=/Td$/] JSXElement[openingElement.name.name=/RadioButton$/][children.length=0]': (node) => {
         context.report({
           node,
           messageId: 'default',
@@ -45,7 +34,7 @@ module.exports = {
           },
         })
       },
-      'JSXElement[openingElement.name.name=/Th$/] JSXElement[openingElement.name.name=/Check(b|B)ox$/]': (node) => {
+      'JSXElement[openingElement.name.name=/Th$/] JSXElement[openingElement.name.name=/Check(b|B)ox$/][children.length=0]': (node) => {
         context.report({
           node,
           messageId: 'default',
@@ -53,14 +42,6 @@ module.exports = {
             cell: 'Th',
             component: 'Checkbox',
             preferred: 'ThCheckbox',
-          },
-          *fix(fixer) {
-            const th = findClosestThFromAncestor(node)
-            if (th) {
-              const thCheckbox = sourceCode.getText(node).replace(/<Check(b|B)ox/, '<ThCheckbox')
-              yield fixer.insertTextAfter(th, thCheckbox)
-              yield fixer.remove(th)
-            }
           },
         })
       },