+++ packages/eslint-plugin-smarthr/rules/trim-props/index.js
@@ -1,5 +1,16 @@
 const SCHEMA = []
-const EXTRA_SPACE_REGES = /(^\s+|\s+$)/
+
+const searchBubbleUp = (node) => {
+  switch (node.type) {
+    case 'Program':
+    case 'JSXAttribute':
+      return null
+    case 'TemplateLiteral':
+      return node
+  }
+
+  return searchBubbleUp(node.parent)
+}
 
 /**
  * @type {import('@typescript-eslint/utils').TSESLint.RuleModule<''>}
@@ -11,23 +22,20 @@ module.exports = {
     fixable: 'whitespace',
   },
   create(context) {
-    return {
-      JSXOpeningElement: (node) =>
-        node.attributes.forEach((current) => {
-          const attribute = current.value?.type === 'JSXExpressionContainer' ? current.value.expression : current.value
-          const props = attribute?.value
-
-          if (typeof props === 'string' && EXTRA_SPACE_REGES.test(props)) {
-            return context.report({
-              node,
-              loc: current.loc,
-              message: '属性に設定している文字列から先頭、末尾の空白文字を削除してください',
-              fix(fixer) {
-                return fixer.replaceTextRange([attribute.range[0] + 1, attribute.range[1] - 1], props.trim())
-              },
-            })
-          }
+    const checker = (node) => {
+      // HINT: TemplateLiteralがネストしている場合、親側だけチェックする
+      if (!searchBubbleUp(node.parent)) {
+        return context.report({
+          node,
+          message: '属性に設定している文字列から先頭、末尾の空白文字を削除してください',
+          fix: (fixer) => fixer.replaceText(node, context.sourceCode.getText(node).replace(/^('|"|`)\s+/, '$1').replace(/\s+('|"|`)$/, '$1')),
         })
+      }
+    }
+
+    return {
+      'JSXAttribute Literal[value=/(^ | $)/]': checker,
+      'JSXAttribute TemplateLiteral:has(>TemplateElement:matches(:first-child[value.raw=/^ /],:last-child[value.raw=/ $/]))': checker,
     }
   },
 }