+++ packages/eslint-plugin-smarthr/test/trim-props.js
@@ -16,8 +16,10 @@ ruleTester.run('trim-props', rule, {
     { code: '<a href="https://www.google.com">google</a>' },
     { code: '<a href={"https://www.google.com"}>google</a>' },
     { code: '<img src="/sample.jpg" alt="sample" />' },
-    { code: '<img src={"/sample.jpg"} alt={"sample"} />' },
+    { code: `<img src={'/sample.jpg'} alt={'sample'} />` },
     { code: '<div data-spec="info-area">....</div>' },
+    { code: '<div data-spec={`a${b} c`}>....</div>' },
+    { code: '<div data-spec={`a${b ? ` ${c} ` : "  "} d`}>....</div>' },
   ],
   invalid: [
     {
@@ -59,8 +61,8 @@ ruleTester.run('trim-props', rule, {
       ],
     },
     {
-      code: '<img src={" /sample.jpg"} alt={"sample "} />',
-      output: '<img src={"/sample.jpg"} alt={"sample"} />',
+      code: `<img src={' /sample.jpg'} alt={'sample '} />`,
+      output: `<img src={'/sample.jpg'} alt={'sample'} />`,
       errors: [
         { message: '属性に設定している文字列から先頭、末尾の空白文字を削除してください' },
         { message: '属性に設定している文字列から先頭、末尾の空白文字を削除してください' },
@@ -74,6 +76,21 @@ ruleTester.run('trim-props', rule, {
     {
       code: '<div data-spec={" info-area "}>....</div>',
       output: '<div data-spec={"info-area"}>....</div>',
+      errors: [{ message: '属性に設定している文字列から先頭、末尾の空白文字を削除してください' }],
+    },
+    {
+      code: '<div data-spec={` ab c `}>....</div>',
+      output: '<div data-spec={`ab c`}>....</div>',
+      errors: [{ message: '属性に設定している文字列から先頭、末尾の空白文字を削除してください' }],
+    },
+    {
+      code: '<div data-spec={` a${b} c `}>....</div>',
+      output: '<div data-spec={`a${b} c`}>....</div>',
+      errors: [{ message: '属性に設定している文字列から先頭、末尾の空白文字を削除してください' }],
+    },
+    {
+      code: '<div data-spec={` a${b ? ` ${c} ` : "  "} d `}>....</div>',
+      output: '<div data-spec={`a${b ? ` ${c} ` : "  "} d`}>....</div>',
       errors: [{ message: '属性に設定している文字列から先頭、末尾の空白文字を削除してください' }],
     },
   ],